fix(dawfile-reaper): tempo is fractional — `as i32` was eating it
`TEMPO 86.28 4 4 0` parsed to 86. The header tempo was typed
`(i32, i32, i32, i32)` and the parser cast through `as i32`, so every
fractional project tempo was silently truncated on the way in — and the
builder truncated the same way on the way out.

That is 9 ms of error per bar at this tempo: two thirds of a second by
the end of a three-minute song, which is well past where a lighting cue
or a click track stops landing. Found while reading a real song project
(*NSYNC, 86.28 BPM) whose regions all sit exactly on bar lines — they
stop doing so the moment the tempo is wrong.

BPM becomes `f64`; the time signature stays integer, because it is. Every
consumer already wrote `bpm as f64`, which now says what it always
meant. `format!("{}", f64)` prints the shortest round-tripping form, so
`TEMPO 120` still writes as `TEMPO 120` and nothing churns.

Test covers 86.28, 128.0001 — four decimal places is what REAPER's tempo
field offers and what a tempo-mapped song contains — and a whole number.

+++ features/dawfile/dawfile-reaper/src/builder.rs
@@ -1471,13 +1471,13 @@ impl ReaperProjectBuilder {
 
     /// Set the project tempo in BPM.
     pub fn tempo(mut self, bpm: f64) -> Self {
-        self.properties.tempo = Some((bpm as i32, 4, 4, 0));
+        self.properties.tempo = Some((bpm, 4, 4, 0));
         self
     }
 
     /// Set the project tempo with a specific time signature.
     pub fn tempo_with_time_sig(mut self, bpm: f64, num: i32, den: i32) -> Self {
-        self.properties.tempo = Some((bpm as i32, num, den, 0));
+        self.properties.tempo = Some((bpm, num, den, 0));
         self
     }
